Don't say "Arabic (Egypt)" in Settings.
Nexus devices only ship one Arabic locale: ar_EG. This should probably be
changed to ar_001 in a later release (and, really, we should let users
choose _all_ of the Arabic-speaking countries), but there's no time for L.

This adds ar_EG to the list of special cases for LocalePicker so that
we'll show "Arabic" instead of "Arabic (Egypt)" (the actual change uses
the Arabic for "Arabic", obviously).

If you're an OEM trying to ship multiple Arabic locales, please accept
my apologies for making your lives slightly harder and, yes, this is the
patch you want to revert so that ar_EG shows up as "Arabic (Egypt)" again.

Bug: 17150708
Change-Id: Iea2fdf813bf3d7758e667bb1a8134a612ee0317b