Fix behavior of inflating InsetDrawable
A recent change to InsetDrawable changed the behavior of inflating
an InsetDrawable when it already had a valid bitmap. The new behavior
avoids throwing an exception with a bad resource where it used to
throw, because the existence of the bitmap makes it avoid trying to
load the resource at all.

The fix is to reintroduce the old behavior of forcing it to load the
resource regardless of the state of its bitmap.

Issue #17068252 InsetDrawable inflation CTS test is failing

Change-Id: I941388730d4479f8f4747a7985754ffdf5133f04