Improve Visibility parent value check.
Bug 17412818
Bug 16398271

The Visibility Transition must not run against children that
are already disappearing or there will be, for example, a double
fade-out effect. Before this change, if a parent matched another
view, it would prevent its child from disappearing.

This change also removes using the overlay for children that have
been removed from the view hierarchy indirectly. This prevents
ListView and RecyclerView children from being added to the overlay.

Change-Id: Iac0610f0939da8643b98812ee1ec1c8d1d70a215