Fix NPE in JobServiceContext
BUG: 17485390
The VERB_BINDING timeout that is set to wait for onBind() to complete
wasn't being cleared when onBind() returns false, i.e. that the service wasn't
available to be bound to.
This led to an NPE when the stale timeout expired. Fix is to clear the timeout
when onBind fails.

Change-Id: I318ca5ce1f3e12b170f7f256608ea7e28f3f120a