Quick fix issue #17514851: dumpsys package <package-name> includes...
...irrelevant install state at the end

Quick and dirty impl just doesn't print any of that data when filtering
by package name.  In the future that part of the dump should be smarter
to know how to filter by package name.  (Probably also moved to a place
earlier in the dump, so the key information -- the overall package
data -- is still at the end.)

Change-Id: I094f7c2f25401438a68a6aa00d10b19c19eb7c7d