Fixing an issue with leaking starting windows. (Bug 17657878)
In some cases, when starting an animation while another starting window is visible,
the starting window is never scheduled to be removed.  In that case, we try and
schedule the closing app starting window to be removed when we are starting the
transition to a new activity.  This also partially addresses issues related to
leaking windows in b/17381033.

Change-Id: Id26525cd71380852f109ec2f55a4a60db5086ded