Fix a bug with screen brightness during screen off animation.
Due to a recent change there was a regression that caused the
screen brightness to be animated down to 0 while the screen
off animation was running.  When the brightness was low this
would cause the screen off animation to be cut short.

This change ensures that we take into account the actual screen
state instead of the desired screen state when making screen
brightness decisions in case we are in the middle of a transition.

The darkness came early.  The pixel fairies trembled.

Bug: 17718416
Change-Id: Ib4b55d61b359abbc70920e324f08a5db07bdd035