Ensure we correctly unblank displays when added.
A previous change introduced a regression in the case where
a device has been added but is initially blanked.  Because
we made changes to defer certain work until we escape the
critical section when making changes to the global display
state, we forgot to apply these changes when displays are
initially added.

This causes problems with HDMI displays remaining blanked
after they are plugged in.

Added a quick fix to ensure we perform the work when adding
a display although we don't bother trying to defer it outside
of the critical section.

Bug: 17909838
Change-Id: If5373d20d8827b7f4330a8cf49f8de64ca3f0740